@@ -10,30 +10,29 @@
     "test": "jest routeErros.test.js --forceExit"
   },
   "dependencies": {
-    "babel-loader": "^8.2.2",
-    "body-parser": "^1.19.0",
-    "chai": "^4.3.4",
+    "babel-loader": "^9.1.0",
+    "body-parser": "^1.20.1",
+    "chai": "^4.3.7",
     "cors": "^2.8.5",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "jest": "^27.1.0",
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2",
+    "jest": "^29.3.1",
     "mysql": "^2.18.1",
-    "nodemon": "^2.0.7",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-global-hook": "^2.0.0-rc2",
-    "styled-components": "^5.3.0",
-    "supertest": "^6.1.6"
+    "nodemon": "^2.0.20",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-global-hook": "^2.0.0",
+    "styled-components": "^5.3.6",
+    "supertest": "^6.3.3"
   },
   "devDependencies": {
-    "@babel/plugin-transform-runtime": "^7.15.0",
-    "@babel/preset-react": "^7.14.5",
+    "@babel/plugin-transform-runtime": "^7.19.6",
+    "@babel/preset-react": "^7.18.6",
     "@types/jest": "^29.2.4",
     "cross-env": "^7.0.3",
-    "html-webpack-plugin": "^5.3.2",
-    "mysql2": "^2.3.3",
-    "webpack": "^5.50.0",
-    "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.11.2"
+    "html-webpack-plugin": "^5.5.0",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.11.1"
   }
 }
