@@ -5,10 +5,12 @@ const route = require('./api/route')
 const cors = require('cors')
 
 
-route.use(cors())
+app.use(cors())
 app.use(express.json())
 app.use(route)
 
 
-
+app.listen(8080,()=>{
+    console.log('rodando')
+})
 module.exports  = app
